validate: Move validate.Report to report.Report
We will use this report in all commands except the test commands, which
must use a different config.
